+++ src/api/index.ts
@@ -1,7 +1,8 @@
 // api-client.ts --------------------------------------------------
 import { apiRequest } from './request';
-import { CreateTalkSessionResponse, Env, TalkSessionInfo } from '../types';
-import { HttpMethod } from '../types';
+import type { Env } from '../shared/types';
+import { TalkSessionInfo } from './types';
+import { HttpMethod } from './types';
 
 export class ApiClient {
   constructor(private env: Env, private jwt?: string) {}
@@ -13,7 +14,7 @@ export class ApiClient {
     });
   }
 
-  getTalk(parentTalkId: string, idempotencyKey: string): Promise<TalkSessionInfo> {
+  createTalkSession(parentTalkId: string, idempotencyKey: string): Promise<TalkSessionInfo> {
     return this.request<TalkSessionInfo>('POST', '/talk/session', {
       parentTalkId,  
       idempotencyKey
