+++ src/shared/errors.ts
