+++ src/shared/types.ts
@@ -1,0 +1,4 @@
+export type Env = {
+  OPENAI_API_KEY: string;
+  API_BASE_URL: string;
+}
